feat: 번개 관련 API 분기 처리 구현

+++ src/service/scrapService.js
@@ -1,8 +1,24 @@
-const { scrapDao } = require('../db');
+const { scrapDao, lightDao, userDao } = require('../db');
+
+const util = require('../lib/util');
+const statusCode = require('../constants/statusCode');
+const responseMessage = require('../constants/responseMessage');
 
 const addLightScrap = async (lightId, memberId) => {
   try {
+    // 존재하는 번개인지 확인
+    const existLight = await lightDao.getExistLight(lightId);
+    if(!existLight) {
+      return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_LIGHT);
+    }
+    // 존재하는 유저인지 확인
+    const exist = await userDao.getUserById(memberId);
+    if (!exist) {
+      return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_USER);
+    }
     await scrapDao.addLightScrap(lightId, memberId);
+    
+    return util.success(statusCode.OK, responseMessage.SCRAP_SUCCESS);
   } catch (error) {
     console.log('postEnterLight error 발생'+ error);
   }
@@ -17,7 +33,14 @@ const getLightScrap = async (lightId, memberId) => {
 }
 const deleteLightScrap = async (lightId, memberId) => {
   try {
+    // 스크랩 한 유저인지 확인
+    const scrapExist = await scrapDao.getLightScrapMember(lightId, memberId);
+    if(!scrapExist){
+      return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_SCRAP);
+    }
     await scrapDao.deleteLightScrap(lightId, memberId);
+    
+    return util.success(statusCode.OK, responseMessage.SCRAP_DELETE_SUCCESS);
   } catch (error) {
     console.log('deleteCancelLight error 발생'+ error);
   }
